requirements.txt
attrs==19.3.0
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
console-menu==0.6.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
Flask==1.1.2
flask-marshmallow==0.12.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
MarkupSafe==1.1.1
marshmallow==3.5.2
matplotlib==3.2.1
numpy==1.18.4
parsel==1.5.2
pkg-resources==0.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
python-dotenv==0.13.0
queuelib==1.5.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.1.0
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
Werkzeug==1.0.1
zope.interface==5.1.0
